Add Wake Lock to Prevent Screen from Turning Off
In the process of sending bulk SMS, wake lock is now acquired to prevent the device screen from turning off. After all messages are sent, wake lock is released. As a result, the device will not go into a sleep mode while the application is sending SMS messages in bulk.